fix terrible missing logic test

+++ sample_work/db.py
@@ -60,8 +60,9 @@ def create_db(c):
     [ c.execute(stmt) for stmt in stmts ]
 
 def make_db(filename, work_items, create):
-    if os.path.exists(filename):
-        os.unlink(filename)
+    if create:
+        if os.path.exists(filename):
+            os.unlink(filename)
 
     database = sqlite.connect(filename)
     c = database.cursor()
